Inherit tests for OpenstackQueryBase
